+++ lib/arel/extensions.rb
@@ -28,6 +28,7 @@ require 'arel/extensions/array_subselect'
 require 'arel/extensions/type_cast'
 require 'arel/extensions/distinct_from'
 require 'arel/extensions/not_distinct_from'
+require 'arel/extensions/null_if'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
