+++ lib/arel/extensions/null_if.rb
@@ -1,0 +1,24 @@
+# rubocop:disable Naming/MethodName
+# rubocop:disable Naming/UncommunicativeMethodParamName
+
+module Arel
+  module Nodes
+    class NullIf < Arel::Nodes::Binary
+    end
+  end
+
+  module Visitors
+    class ToSql
+      def visit_Arel_Nodes_NullIf(o, collector)
+        collector << 'NULLIF('
+        visit o.left, collector
+        collector << ', '
+        visit o.right, collector
+        collector << ')'
+      end
+    end
+  end
+end
+
+# rubocop:enable Naming/MethodName
+# rubocop:enable Naming/UncommunicativeMethodParamName
